Ecriture indexpointfile ok pour las et laz
git-svn-id: http://rdinnovation.onf.fr/svn/computreev6@342 b135c390-378b-4ba5-bdcd-e4540112eb5a

+++ library/ctliblas/readers/ct_reader_lasv2.cpp
@@ -892,7 +892,7 @@ bool CT_Reader_LASV2::internalReadMultiFile(CT_StandardItemGroup *group)
 
 
 
-bool CT_Reader_LASV2::getPointIndicesInside2DShape(const CT_AreaShape2DData* area2D, size_t &lastIncludedIndex, QList<size_t> &indicesAfterLastIncludedIndex) const
+bool CT_Reader_LASV2::getPointIndicesInside2DShape(const CT_AreaShape2DData* area2D, qint64 &lastIncludedIndex, QList<size_t> &indicesAfterLastIncludedIndex) const
 {
     bool all = true;
     lastIncludedIndex = -1;
